Add cvars to reward ubercharges and building upgrades.
Building upgrades are not rewarded by default, but can be run-time enabled to.
Both of these are minorly abusable in that they can be performed during setup,
so if they're rewarded at all, they should be kept very low scoring - in fact,
building upgrades should probably reward less than 10% of the requirement (not
possible with the default values, so sm_buffbot_carnage_required would need to
be increased). Use at your own risk.