mvm: Rework how money is specified.
Instead of assigning money per subwave, the code now sets money per enemy, and
it's then multiplied by the enemy count (including for support waves; so the
first ten will have the specified money, and then there'll be none). This
entailed a reworking of the money drops per wave, and now you get a lot more
in the fifth wave. The early waves are approximately comparable to their old
drops, though.