Bugfix teleportation...
Previously, if the trace attempts all failed at the very start, the target
location would not be updated, and you would teleport to the map origin. A
cool feature, to be sure, but not exactly the intention... especially since
you could potentially glitch your way to an impossible location, or (as in
the case of pl_upward) a place that's going to instakill you. Now fixed.