Rework the weapon name translation table.
Instead of deeming some weapons "boring", we now check for what's the current
default primary/secondary for each team, and consider THOSE boring (in terms
of weapon drops). So if you have a game mode where everyone starts with an
AWP, you'll see no messages when someone drops an AWP in favour of an XM1014.

This then allows other parts of the code - such as health gating - to use the
same translation table while counting all weapons as "interesting".