Change respawn lag immunity.
Instead of using deathmatch immunity, it instead just flags you as not
receiving damage events, and then everything else is done here. The result
is a Coyote-style smoky face effect (or Lidda from D&D, if you prefer),
and a greyed-out view for a while. This should deal with the previous
issues regarding warmup and the time not being reliable or correct.